Addressed few comments
